+++ lua/plugins/nvim-spectre.lua
@@ -1,0 +1,10 @@
+-- lua/plugins/nvim-spectre.lua
+
+return {
+    'nvim-pack/nvim-spectre',
+    dependencies = { 'nvim-lua/plenary.nvim' },
+    cmd    = 'Spectre',
+    config = function()
+        require('spectre').setup()
+    end,
+}