+++ lua/plugins/zen-mode.lua
@@ -1,0 +1,9 @@
+-- lua/plugins/zen-mode.lua
+
+return {
+    'folke/zen-mode.nvim',
+    cmd    = 'ZenMode',
+    config = function()
+        require('zen-mode').setup()
+    end,
+}